requirements.txt
fastapi
uvicorn
requests
serpapi
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
bs4
